feat: make robots author code to cleanup robots code
