+++ src/components/Layout/Header/NavBar/DrawerWalletHeader.tsx
@@ -15,9 +15,7 @@ import type { FC } from 'react'
 import { memo, useCallback, useMemo } from 'react'
 import { TbDots } from 'react-icons/tb'
 import { useTranslate } from 'react-polyglot'
-import { useNavigate } from 'react-router'
 
-import { DrawerWalletRoutes } from './DrawerWalletSettings/DrawerSettingsRoutes'
 import { WalletImage } from './WalletImage'
 
 import { SUPPORTED_WALLETS } from '@/context/WalletProvider/config'
@@ -38,13 +36,13 @@ type DrawerHeaderProps = {
   onDisconnect: () => void
   onSwitchProvider: () => void
   onClose?: () => void
+  onSettingsClick?: () => void
 }
 
 export const DrawerWalletHeader: FC<DrawerHeaderProps> = memo(
-  ({ walletInfo, isConnected, connectedType, onDisconnect, onSwitchProvider, onClose }) => {
+  ({ walletInfo, isConnected, connectedType, onDisconnect, onSwitchProvider, onSettingsClick }) => {
     const translate = useTranslate()
     const settings = useModal('settings')
-    const navigate = useNavigate()
 
     const maybeRdns = useAppSelector(selectWalletRdns)
     const mipdProviders = useMipdProviders()
@@ -59,19 +57,12 @@ export const DrawerWalletHeader: FC<DrawerHeaderProps> = memo(
     )
 
     const handleSettingsClick = useCallback(() => {
-      // If we're in a drawer context (have onClose prop), use drawer routing
-      if (onClose) {
-        try {
-          navigate(DrawerWalletRoutes.Settings)
-        } catch {
-          // If navigate fails (not in router context), fallback to modal
-          settings.open({})
-        }
+      if (onSettingsClick) {
+        onSettingsClick()
       } else {
-        // Otherwise, use the modal
         settings.open({})
       }
-    }, [settings, navigate, onClose])
+    }, [settings, onSettingsClick])
 
     const repeatIcon = useMemo(() => <RepeatIcon />, [])
     const closeIcon = useMemo(() => <CloseIcon />, [])
