+++ src/components/Modals/Settings/CurrencyFormat.tsx
@@ -1,5 +1,5 @@
 import { ArrowBackIcon } from '@chakra-ui/icons'
-import { Button, Flex, Icon, IconButton, ModalBody, ModalHeader } from '@chakra-ui/react'
+import { Button, Flex, Icon, IconButton, ModalBody, ModalHeader, Stack } from '@chakra-ui/react'
 import sortBy from 'lodash/sortBy'
 import { useCallback } from 'react'
 import { FaCheck } from 'react-icons/fa'
@@ -15,7 +15,11 @@ import { useAppDispatch, useAppSelector } from '@/state/store'
 
 const arrowBackIcon = <ArrowBackIcon />
 
-export const CurrencyFormat = () => {
+type CurrencyFormatProps = {
+  isDrawer?: boolean
+}
+
+export const CurrencyFormat = ({ isDrawer = false }: CurrencyFormatProps) => {
   const dispatch = useAppDispatch()
   const currentCurrencyFormat = useAppSelector(preferences.selectors.selectCurrencyFormat)
   const selectedCurrency = useAppSelector(preferences.selectors.selectSelectedCurrency)
@@ -29,6 +33,36 @@ export const CurrencyFormat = () => {
   const handleGoBack = useCallback(() => {
     navigate(-1)
   }, [navigate])
+
+  if (isDrawer) {
+    return (
+      <Stack width='full' p={0} spacing={2}>
+        {formats.map(currencyFormat => {
+          const active = currencyFormat === currentCurrencyFormat
+          const buttonProps = active
+            ? {
+                isDisabled: true,
+                _disabled: { opacity: 1 },
+              }
+            : {
+                pl: 8,
+                variant: 'ghost',
+                onClick: () => dispatch(setCurrencyFormat({ currencyFormat })),
+              }
+          return (
+            <Button width='full' justifyContent='flexStart' key={currencyFormat} {...buttonProps}>
+              <Flex alignItems='center' textAlign='left'>
+                {active && <Icon as={FaCheck} color='blue.500' />}
+                <Flex ml={4}>
+                  <RawText>{currencyFormatsRepresenter(currencyFormat, selectedCurrency)}</RawText>
+                </Flex>
+              </Flex>
+            </Button>
+          )
+        })}
+      </Stack>
+    )
+  }
 
   return (
     <SlideTransition>
