+++ src/components/Modals/Settings/FiatCurrencies.tsx
@@ -1,5 +1,5 @@
 import { ArrowBackIcon } from '@chakra-ui/icons'
-import { Button, Flex, Icon, IconButton, ModalBody, ModalHeader } from '@chakra-ui/react'
+import { Button, Flex, Icon, IconButton, ModalBody, ModalHeader, Stack } from '@chakra-ui/react'
 import identity from 'lodash/identity'
 import sortBy from 'lodash/sortBy'
 import { useCallback } from 'react'
@@ -16,7 +16,11 @@ import { useAppDispatch, useAppSelector } from '@/state/store'
 
 const arrowBackIcon = <ArrowBackIcon />
 
-export const FiatCurrencies = () => {
+type FiatCurrenciesProps = {
+  isDrawer?: boolean
+}
+
+export const FiatCurrencies = ({ isDrawer = false }: FiatCurrenciesProps) => {
   const dispatch = useAppDispatch()
   const selectedCurrency = useAppSelector(preferences.selectors.selectSelectedCurrency)
   const translate = useTranslate()
@@ -31,6 +35,38 @@ export const FiatCurrencies = () => {
   const handleGoBack = useCallback(() => {
     navigate(-1)
   }, [navigate])
+
+  if (isDrawer) {
+    return (
+      <Stack width='full' p={0} spacing={2}>
+        {allFiatCurrencies.map(currency => {
+          const active = currency === selectedCurrency
+          const buttonProps = active
+            ? {
+                isDisabled: true,
+                _disabled: { opacity: 1 },
+              }
+            : {
+                pl: 8,
+                variant: 'ghost',
+                onClick: () => dispatch(setSelectedCurrency({ currency })),
+              }
+          return (
+            <Button width='full' justifyContent='flexStart' key={currency} {...buttonProps}>
+              <Flex alignItems='center' textAlign='left'>
+                {active && <Icon as={FaCheck} color='blue.500' />}
+                <Flex ml={4}>
+                  <RawText>{currency}</RawText>
+                  <RawText mx={2}>-</RawText>
+                  <Text translation={`modals.settings.currencies.${currency}`} />
+                </Flex>
+              </Flex>
+            </Button>
+          )
+        })}
+      </Stack>
+    )
+  }
 
   return (
     <SlideTransition>
